requirements.txt
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
google-api-core==2.0.1
google-auth==2.2.1
google-cloud-core==2.0.0
google-cloud-firestore==2.3.4
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna==3.2
packaging==21.0
proto-plus==1.19.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
requests==2.26.0
retrying==1.3.3
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
beautifulsoup4~=4.10.0